objformat: don't hardcode compiler versions
In the previous commit a fallback to the "custom" compiler driver was
introduced for unknown compilers.  The previous approach hard-coded
known compiler versions, making it more complicated to add new compilers
to base.

This commit changes slightly the behavior for the "custom" compiler
fallback:  The compiler is searched for like before the "custom" change,
and in addition, in case no compiler could be found, the search is
retried for the "custom" compiler driver.  This way we don't have to
hard-code known compiler versions, while a truely non-existing CCVER
will still be handled by the "custom" compiler driver.