bsd.*.mk: allow the setting of separate compile flags for files
Our makefile infrastructure assumed that all compile flags for all files
are the same.  Now you can specify compile flags for file groups, much
like for FILES and INCS.

This does not support files with pathname components yet.  For this to
work, we'd need support in make to retrieve the full seached path of a
file.