kernel module loader - fix load dependencies from preload modules by kernel
The boot loader is supposed to load module dependencies but it misses
a few when e.g. 'snd_hda' is specified in loader.conf.  This resulted
in the kernel itself trying to load the missing dependencies, but the
kernel was running the sysinits for those dependencies immediately during
the preload stage instead of registering them.

Pass a flag to the loader chain to register the sysinits instead of
running them if a dependency is loaded during the preload stage.  This
fixes the problem.  And also, by the way, the loader does not need to
load dependencies any more (though it still does).

Reported-by: corecode, swildner