DEVFS - Use v_associate_rdev; fix vcount/count_dev
to avoid trouble with both vrevoke() and the drm code which calls
vm_mmap, which in turn expects a vnode as handle, but drm only has
a cdev. If associate_rdev is used, the cdev will have a list of
vnodes currently in use  associated with it.

- This fixes a problem with drm which caused a panic when used.

- count_dev, which is used by vcount, now returns the vnodes'
opencount instead of the number of references.

Reported-By: Johannes Hofmann