hexdump: always stat input when skipping
The previous code would not fstat(2) the input, if operating on stdin.
Still, it would use the uninitialized struct stat to decide whether to
seek in the input or not.  This could either lead to hexdump reading the
skipped input, instead of directly seeking, or, worse, trying a seek
even though this not being possible.

Fix this issue by always fstat()ing the input.

Submitted-by: corecode