buildworld: properly sequence quickworld
We must not run any of build stage in parallel.  The existing .ORDER
only works for the sequence of a full buildworld.  It seems that make is
not ordering any pair in an .ORDER target, but only pairs as written.
Quickworld does not run the intermediate stages between _obj and
_includes, thus allowing make to parallelize these two.

The proper fix might be in make, if we want to change the semantics of
the .ORDER target, but fixing it here doesn't hurt in any case.