Fix registration issue in the devfs rc script.
rc.d/devfs is not controlled by an _enable variable so just register properly
as being configured, so that later scripts indirectly depending on it will
not fail when being started by rcrun(8).

Reported-by: hasso

While here, add some comments about the scripts purpose and print rules files
as they are applied.