top(1): Silence gcc warnings.
The ORDERKEY_XXX() macros build up a deeply nested set of if-statements.
E.g.,

if (expr)
   if (expr)
      if (expr)
      	 if (expr)

The last if-statement needs to contain nothing (every expression has
its own side effect, so the whole cascade is actually meaningful). Using
a ; though, makes gcc generate warnings regarding 'empty body in if
statements'. That's why we, instead, use:

ORDERKEY_FOO1
ORDERKEY_FOO2
...
ORDERKEY_FOO3
{}

We could have embedded the {} in the definition of the macro itself, but
that would break if someone rearranged the macros' call order.

Reviewed-by: dillon@