Kernel - change sync_on_panic default to 0
No longer attempt to sync the buffer cache when a panic occurs.  This
has always been problematic, HAMMER will ignore the attempt anyway,
and now we are formalizing that we no longer try.

This also improves the chances of the system dumping and rebooting
after a panic vs freezing up.