installer: More /dev/serno work.
Take back the devtab(5) support in the installer.

mountroot didn't work for the UFS with serial numbers case since the
data for '/' is taken from fstab and that would have a devtab label
in it and we're not setting vfs.root.mountfrom explicitly for UFS.
We don't want to, either.

So just use /dev/serno/* nodes (which will work everywhere) always.

Suggested-by: dillon