fdisk - Use heads = 255 on file images
When operating on a file image use max values for sectors and heads in
order to match the case where the file might exceed the CHS limit.  By
convention the CHS fields are set to all 1's in this case.

If our sectors and heads are not set at max values then the all 1's case
will result in a geometry that overflows the originally specified sectors
and heads which can brick the BIOS.