core dumps: don't claim we are FreeBSD
Our elf core dump format (slightly) changed before, and we could only
convince gdb to work nicely by changing gdb's FreeBSD support.  Now fix
it the right way by not claiming that we are FreeBSD.

At the same time, change the way tids are stored in the elf notes.  This
is actually like FreeBSD does it.