mqueues: Let the world know we support them.
There are two ways to determine whether we support the message
passing interface, as described in POSIX 1003.1B standard.

The standard way is to use sysconf(3) and the DragonFly-specific
way is to check the value of p1003_1b.message_passing sysctl.

We make sure that both return consistent results with respect to
each other.