evtranalyze: add summary command and refactor interval code
Allow the user to specify a time interval in miliseconds. Coupled
with the summary command, this makes it easy to select a subset of
the events for the svg command.

Also fix the svg code to not assume there are always 5 threads
with activity.

Add a new option to the svg command to write to a different path
(the default remains "output.svg").