nrelease: Remove LiveDVD specific kernel configs and loader menu.
These were introduced in order to also offer SMP kernels (with and
without APIC_IO) from the LiveDVD. But we don't want to maintain 4
extra kernel configs just for that purpose, especially since we
have x86_64 support now which would require introducing even more
configs. The whole concept has to be re-thought.

So until we come up with a better solution, get rid of it and go
with the usual GENERIC/VKERNEL (X86_64_GENERIC/VKERNEL64 on
x86_64) combination just like we do for the LiveCD.