sys/netinet/in_pcb.c: Fix a broken range check.
Back in 2004, in commit 80dff869, RANGECHK() was changed from a macro
to an inline function. This broke the test since the passed 'var' could
not be changed by the inline unless a pointer was used.

Instead of changing to using pointers now (or adding a return value to
RANGECHK()), revert to using a macro again.