vesa(4): Remove definition and usage of the obsolete V_MODEOPTINFO flag.
Prior to VBE 1.2, this flag specified whether optional information
for the mode was available in the ModeInfoBlock structure. Starting
with 1.2 this information became mandatory and the D1 bit became
reserved.

Since we claim to only support cards conforming to VBE 1.2 and higher,
adjust the code to no longer take this bit into account and remove its
definition.