math.h: Convert HUGE_VALL to constant, from const-qualified var
Define HUGE_VALL via a compiler builtin. If the compiler doesn't
support it, we just fallback to previous situation. Both gcc and
clang play well with it.

This patch allows us to:
1. use HUGE_VALL as initializer element for objects with static
   storage duration
2. not worry about the format of long double (e.g., 96bit vs 128bit)
3. not worry about things like endianness

Briefly-discussed-with: corecode@, swildner@, but put the blame on me
if I broke something for you.

Special-thanks-to: swildner@ for checking code with clang.

P.S. HUGE_VAL{,F} already do the right thing.