Make the loader load ehci.ko by default.
It seems to do more good than harm.

HOWEVER!

Unfortunately, There are systems that don't play nice with ehci. One
of my boxes, for example, won't reboot with ehci loaded.

Therefore, the behavior introduced with this commit can be disabled
by doing one of the following things:

1) 'unset ehci_load' from the loader prompt before booting.

2) Booting in 'safe mode'.

3) Putting 'hint.ehci.0.disabled=1' in /boot/loader.conf.

This is all modeled after the way the automatic loading of acpi.ko is
handled.