gzip - Do not return from write after each inflate
From FreeBSDs commit message:

In the past gunzip(1) write()'s after each inflate return.  This is
not optimal from a performance standpoint since the write buffer is
not necessarily be filled up when the inflate rountine reached the
end of input buffer and it's not the end of file.

Submitted-By: Xin LI <delphij@FreeBSD.org>
Obatined-From: FreeBSD