installer: Raise the recommended size for / in the UFS install.
It turned out that the x86_64 install's contents of / didn't fit into
256M. So increase to 320M for the smallest supported disks (2-4G) and
raise it gradually to 640M (4-10G) and 768 (anything above 10G). The
latter is also the HAMMER install's default recommendation.

The best solution here would probably be to base the recommendations
on the actual contents of the medium (to be tackled after the release).

Reported-by: Przemyslaw Pawelczyk <pp_o2@o2.pl>