Bring in mps(4) for LSI Fusion-MPT 2 Serial Attached SCSI controllers.
The driver should support the following controllers:

* LSI Logic SAS2004 (4 Port SAS)
* LSI Logic SAS2008 (8 Port SAS)
* LSI Logic SAS2108 (8 Port SAS)
* LSI Logic SAS2116 (16 Port SAS)
* LSI Logic SAS2208 (8 Port SAS)

Due to it still being in in development (Integrated RAID isn't supported,
for example), it's only hooked into the module build and added to LINT.

The port hasn't received any testing at all other than make it build. But
it is known that Matt has such a controller. :-)

Thanks to FreeBSD from which this driver is taken.