bluetooth(4): Fix loading bluetooth without pf.
bluetooth(4) was previously using pf(4)'s pool_* macros. Now that pool_get()
has been turned into a function in pf(4), this is no longer possible, since
the bluetooth(4) module would have to depend on the pf(4) module for that.

To unbreak module loading, convert all calls to these macros to the zone(9)
calls which they really are.